Refactor: nfa_dfa_core.py to core.py nfa_dfa_graph.py to graph.py nfa_dfa_latex.py to latex.py
